/* PR_START p--query-resolution-perf 05 */ Move filter-related classes to a separate module.
