+++ metricflow-semantics/metricflow_semantics/specs/where_filter/where_filter_time_dimension.py
@@ -23,7 +23,7 @@ from metricflow_semantics.query.group_by_item.filter_spec_resolution.filter_spec
 )
 from metricflow_semantics.specs.column_assoc import ColumnAssociationResolver
 from metricflow_semantics.specs.rendered_spec_tracker import RenderedSpecTracker
-from metricflow_semantics.specs.where_filter_dimension import WhereFilterDimension
+from metricflow_semantics.specs.where_filter.where_filter_dimension import WhereFilterDimension
 
 
 class WhereFilterTimeDimension(WhereFilterDimension):
