+++ metricflow-semantics/metricflow_semantics/specs/where_filter/where_filter_transform.py
@@ -15,10 +15,10 @@ from metricflow_semantics.query.group_by_item.filter_spec_resolution.filter_spec
 from metricflow_semantics.specs.column_assoc import ColumnAssociationResolver
 from metricflow_semantics.specs.rendered_spec_tracker import RenderedSpecTracker
 from metricflow_semantics.specs.spec_classes import WhereFilterSpec
-from metricflow_semantics.specs.where_filter_dimension import WhereFilterDimensionFactory
-from metricflow_semantics.specs.where_filter_entity import WhereFilterEntityFactory
-from metricflow_semantics.specs.where_filter_metric import WhereFilterMetricFactory
-from metricflow_semantics.specs.where_filter_time_dimension import WhereFilterTimeDimensionFactory
+from metricflow_semantics.specs.where_filter.where_filter_dimension import WhereFilterDimensionFactory
+from metricflow_semantics.specs.where_filter.where_filter_entity import WhereFilterEntityFactory
+from metricflow_semantics.specs.where_filter.where_filter_metric import WhereFilterMetricFactory
+from metricflow_semantics.specs.where_filter.where_filter_time_dimension import WhereFilterTimeDimensionFactory
 from metricflow_semantics.sql.sql_bind_parameters import SqlBindParameters
 
 logger = logging.getLogger(__name__)
